organizational refactor
